requirements.txt
fastapi>=0.111,<0.112
uvicorn[standard]>=0.25,<0.26
sqlalchemy>=2.0.48,<2.1
psycopg2-binary>=2.9.7,<3
pydantic>=2.3,<2.4
python-jose[cryptography]>=3.3,<3.4
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv>=1.0,<1.1
python-multipart>=0.0.7
werkzeug==3.0.1

# Для обработки PDF
pdfplumber==0.10.3

# Для определения языка
langdetect==1.0.9

# Для AI (опционально, можно закомментировать если не используется)
openai>=1.0.0,<1.8
httpx>=0.24.1,<0.25

# Для миграций БД
alembic>=1.12.0
python-multipart